web/regular-forum-page/src/web/src/regularforum/requirements.txt
asgiref==3.3.1
Django==3.1.5
dj-database-url==0.5.0
django-ratelimit==3.0.1
psycopg2-binary>=2.8.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0